+++ moirae/estimators/offline/extractors/base.py
