+++ moirae/estimators/offline/extractors/ecm/capacity.py
@@ -4,7 +4,7 @@ Defines capacity extractor
 from battdat.data import BatteryDataset
 from battdat.postprocess.integral import CapacityPerCycle
 
-from moirae.extractors.base import BaseExtractor
+from moirae.estimators.offline.extractors.base import BaseExtractor
 from moirae.models.ecm.components import MaxTheoreticalCapacity
