+++ tests/estimators/offline/extractors/test_ecm_rc.py
@@ -3,7 +3,7 @@ from pytest import fixture, raises
 import numpy as np
 
 from battdat.postprocess.integral import StateOfCharge
-from moirae.extractors.ecm import MaxCapacityExtractor, RCExtractor
+from moirae.estimators.offline.extractors.ecm import MaxCapacityExtractor, RCExtractor
 from moirae.models.ecm.utils import unrealistic_fake_rc
